server/requirements.txt
fastapi
uvicorn
psycopg2-binary
mercadopago
requests
python-dotenv
